some suggested changes done for improving route file
