+++ src/V1/Controllers/PaypalController.php
@@ -2,16 +2,9 @@
 
 namespace Immera\Payment\V1\Controllers;
 
-use App\Http\Controllers\Controller;
-use Exception;
-use Illuminate\Http\Request;
-use Immera\Payment\Events\PaymentInstanceCreated;
-use Immera\Payment\Events\PaymentInstanceUpdated;
-use Immera\Payment\Models\PaymentInstance;
-use Immera\Payment\Payment;
 use Immera\Payment\Controllers\PaypalController as PaypalControllerDefault;
 
 class PaypalController extends PaypalControllerDefault
 {
-    
+    // will over ride methods as needed.
 }
